Use token

+++ coupon_system_spring/src/main/java/app/core/controllers/AdminController.java
@@ -10,44 +10,30 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import app.core.entities.Company;
 import app.core.entities.Customer;
 import app.core.exceptions.CouponSystemException;
-import app.core.login.ClientType;
-import app.core.login.LoginManager;
 import app.core.services.AdminService;
 
 @RestController
 @RequestMapping("/api")
-public class AdminController implements ClientController {
-
+public class AdminController {
+	
+	@Autowired
 	private AdminService service;
 	@Autowired
-	private LoginManager loginManager;
-	
-	@Override
-	@PostMapping("/login/{email}/{password}")
-	public boolean login(@PathVariable String email, @PathVariable String password) {
-		System.out.println("AdminController login");
-		try {
-			service = (AdminService) loginManager.login(email, password, ClientType.ADMINISTRATOR);
-			if (service != null) {
-				return true;
-			}
-			throw new CouponSystemException(HttpStatus.BAD_REQUEST, "login fail :(");
-		} catch (CouponSystemException e) {
-			throw e;
-		} catch (Exception e) {
-			throw new CouponSystemException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage(), e);
-		}
-	}
+	private TokenValidator tokenValidator;
 	
 	@PostMapping("/add-company")
-	public Company addCoumpany(@RequestBody Company company) {
+	public Company addCoumpany(@RequestBody Company company, @RequestHeader String token) {
 		System.out.println("AdminController addCompany");
+		if (tokenValidator.validate(token)) {
+			throw new CouponSystemException(HttpStatus.UNAUTHORIZED, "You need to login");
+		}
 		try {
 			Company addedCompany = service.addCompany(company);
 			return addedCompany;
@@ -59,8 +45,11 @@ public class AdminController implements ClientController {
 	}
 	
 	@PutMapping("/update-company")
-	public Company updateCompany(@RequestBody Company company) {
+	public Company updateCompany(@RequestBody Company company, @RequestHeader String token) {
 		System.out.println("AdminController updateCompany");
+		if (tokenValidator.validate(token)) {
+			throw new CouponSystemException(HttpStatus.UNAUTHORIZED, "You need to login");
+		}
 		try {
 			Company updatedCompany = service.updateCompany(company);
 			return updatedCompany;
@@ -73,7 +62,11 @@ public class AdminController implements ClientController {
 	}
 	
 	@DeleteMapping("/delete-company/{id}")
-	public Company deleteCompany(@PathVariable Long id) {
+	public Company deleteCompany(@PathVariable Long id, @RequestHeader String token) {
+		System.out.println("AdminController deleteCompany");
+		if (tokenValidator.validate(token)) {
+			throw new CouponSystemException(HttpStatus.UNAUTHORIZED, "You need to login");
+		}
 		try {
 			Company  deletedCompany = service.deleteCoumpany(id);
 			return deletedCompany;
@@ -85,8 +78,11 @@ public class AdminController implements ClientController {
 	}
 	
 	@GetMapping("/companies")
-	public List<Company> getAllCompanies() {
+	public List<Company> getAllCompanies(@RequestHeader String token) {
 		System.out.println("AdminController getAllCompanies");
+		if (tokenValidator.validate(token)) {
+			throw new CouponSystemException(HttpStatus.UNAUTHORIZED, "You need to login");
+		}
 		try {
 			List<Company> companies = service.getAllCompanies();
 			return companies;
@@ -99,8 +95,11 @@ public class AdminController implements ClientController {
 	}
 	
 	@GetMapping("/company/{id}")
-	public Company getCompany(@PathVariable Long id) {
+	public Company getCompany(@PathVariable Long id, @RequestHeader String token) {
 		System.out.println("AdminController getCompany");
+		if (tokenValidator.validate(token)) {
+			throw new CouponSystemException(HttpStatus.UNAUTHORIZED, "You need to login");
+		}
 		try {
 			Company compnay = service.getCompany(id);
 			return compnay;
@@ -112,8 +111,11 @@ public class AdminController implements ClientController {
 	}
 	
 	@GetMapping("/company/name/{name}")
-	public Company getCompanyByName(@PathVariable String name) throws CouponSystemException {
+	public Company getCompanyByName(@PathVariable String name, @RequestHeader String token) throws CouponSystemException {
 		System.out.println("AdminController getCompanyByName");
+		if (tokenValidator.validate(token)) {
+			throw new CouponSystemException(HttpStatus.UNAUTHORIZED, "You need to login");
+		}
 		try {
 			Company company = service.getCompanyByName(name);
 			return company;
@@ -127,8 +129,11 @@ public class AdminController implements ClientController {
 	// ********** Customer Methods ***************** //
 	
 	@PostMapping("/add-customer")
-	public Customer addCustomer(@RequestBody Customer customer) {
+	public Customer addCustomer(@RequestBody Customer customer, @RequestHeader String token) {
 		System.out.println("AdminController addCustomer");
+		if (tokenValidator.validate(token)) {
+			throw new CouponSystemException(HttpStatus.UNAUTHORIZED, "You need to login");
+		}
 		try {
 			Customer addedCustomer = service.addCustomer(customer);
 			return addedCustomer;
@@ -141,8 +146,11 @@ public class AdminController implements ClientController {
 	}
 	
 	@PutMapping("/update-customer")
-	public Customer updateCustomer(@RequestBody Customer customer) {
+	public Customer updateCustomer(@RequestBody Customer customer, @RequestHeader String token) {
 		System.out.println("AdminController updateCustomer");
+		if (tokenValidator.validate(token)) {
+			throw new CouponSystemException(HttpStatus.UNAUTHORIZED, "You need to login");
+		}
 		try {
 			Customer updatedCustomer = service.updateCustomer(customer);
 			return updatedCustomer;
@@ -155,8 +163,11 @@ public class AdminController implements ClientController {
 	}
 	
 	@DeleteMapping("/delete-customer/{id}")
-	public Customer deleteCustomer(@PathVariable Long id) {
+	public Customer deleteCustomer(@PathVariable Long id, @RequestHeader String token) {
 		System.out.println("AdminController deleteCustomer");
+		if (tokenValidator.validate(token)) {
+			throw new CouponSystemException(HttpStatus.UNAUTHORIZED, "You need to login");
+		}
 		try {
 			Customer deletedCustomer = service.deleteCustomer(id);
 			return deletedCustomer;
@@ -169,8 +180,11 @@ public class AdminController implements ClientController {
 	}
 	
 	@GetMapping("/customers")
-	public List<Customer>  getAllCustomers() {
+	public List<Customer>  getAllCustomers(@RequestHeader String token) {
 		System.out.println("AdminController getAllCustomers");
+		if (tokenValidator.validate(token)) {
+			throw new CouponSystemException(HttpStatus.UNAUTHORIZED, "You need to login");
+		}
 		try {
 			List<Customer> customers = service.getAllCustomer();
 			return customers;
@@ -182,8 +196,11 @@ public class AdminController implements ClientController {
 	}
 	
 	@GetMapping("/customer/{id}")
-	public Customer getCustomer(@PathVariable Long id) {
+	public Customer getCustomer(@PathVariable Long id, @RequestHeader String token) {
 		System.out.println("AdminController getCustomer");
+		if (tokenValidator.validate(token)) {
+			throw new CouponSystemException(HttpStatus.UNAUTHORIZED, "You need to login");
+		}
 		try {
 			Customer customer = service.getCustomer(id);
 			return customer;
